feat: first state vectors from SGP4

+++ src/libraries/c/external/sgp4/TLE.h
@@ -27,7 +27,7 @@ typedef struct TLE {
 
 void parseLines(TLE *tle, char *line1, char *line2, int wgs_model);
 
-long parseEpoch(ElsetRec *rec, char *str);
+double parseEpoch(ElsetRec *rec, char *str);
 
 void getRVForDate(TLE *tle, long millisSince1970, double r[3], double v[3]);
